Lifting state up

+++ posts/src/Posts.js
@@ -1,19 +1,17 @@
-import { useEffect, useState } from "react";
 import axios from 'axios';
 import { Button, Table } from 'react-bootstrap';
 import { XOctagon } from 'react-bootstrap-icons'
 
-function Posts() {
-    const [posts, setPosts] = useState([]);
-    useEffect(() => {
-        axios.get('http://localhost:3000/posts')
-            .then(({ data }) => {
-                setPosts(data);
-            })
-    }, []);
-    const remove = async (id) => {
-        await axios.delete(`http://localhost:3000/posts/${id}`);
-        alert('Post removed!');
+function Posts({ data = [], getDataFromServer }) {
+
+
+    const deletePost = async (id) => {
+        const answer = window.confirm(`Are you sure you want to delete post ${id}?`);
+        if (answer) {
+            await axios.delete(`http://localhost:3000/posts/${id}`);
+            alert(`Post ${id} has been deleted!`);
+            getDataFromServer();
+        }
     }
 
     return (
@@ -24,19 +22,25 @@ function Posts() {
                     <th>id</th>
                     <th>title</th>
                     <th>body</th>
-                    <th>remove</th>
+                    <th>Delete</th>
                 </tr>
             </thead>
             <tbody>
                 {
-                    posts.map((posts) => (
+                    data.map((posts) => (
                         <tr>
                             <td>{posts.userId}</td>
                             <td>{posts.id}</td>
                             <td>{posts.title}</td>
                             <td>{posts.body}</td>
                             <div>
-                                <Button size="lg" variant="link" onClick={() => remove(posts.id)}><XOctagon className="text-danger"></XOctagon></Button>
+                                <Button
+                                    size="lg"
+                                    variant="link"
+                                    onClick={() => deletePost(posts.id)}
+                                >
+                                    <XOctagon className="text-danger"></XOctagon>
+                                </Button>
                             </div>
                         </tr>
                     ))
